Moved InstanceConfigRpcCache to v1 package per comment Changed to package private
